% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_detection.R
\name{pattern_detection}
\alias{pattern_detection}
\title{Pattern Detection}
\usage{
pattern_detection(
  indicator,
  group,
  period_end,
  numerator,
  denominator,
  multiplier,
  better_is,
  overdispersion = FALSE,
  spc_chart_type = "p",
  funnel_chart_type = "PR",
  parent_indicator = NA,
  indicator_name = indicator,
  group_name = group,
  funnel_data_points = "Yes",
  trend_size = 5,
  shift_size = 7
)
}
\arguments{
  \item{indicator}{A vector of indicator codes.}

  \item{group}{A vector of group IDs.}

  \item{period_end}{A vector of dates. Format: \%d/\%m/\%Y (e.g., 12/11/2021).}

  \item{numerator}{A numeric vector.}

  \item{denominator}{A numeric vector.}

  \item{multiplier}{A numeric value used as a multiplication factor. Default is 1.}

  \item{better_is}{A character string indicating the direction of positive change ("higher" or "lower").}

  \item{overdispersion}{Logical. If TRUE, adjusts for overdispersion in funnel plot calculations. Default is FALSE.}

  \item{spc_chart_type}{A string identifying the type of SPC chart. Default is "p".}

  \item{funnel_chart_type}{A string identifying the type of funnel plot. Default is "PR".}

  \item{parent_indicator}{A vector used for indicator grouping.}

  \item{indicator_name}{A vector of descriptive names for indicators.}

  \item{group_name}{A vector of descriptive names for groups.}

  \item{funnel_data_points}{A vector indicating which data points are included in the funnel plot calculation.}

  \item{trend_size}{The number of points in a trend pattern.}

  \item{shift_size}{The number of points in a shift pattern.}
}
\value{
A data table of unique indicator-group rows with any identified patterns.
}
\description{
Pattern Detection
}
\seealso{
Other Pattern detection functions: 
\code{\link{pattern_rules}()},
\code{\link{valid_spc}()}
}
\concept{Pattern detection functions}
